WIP - Feature de adicionar ao carrinho
Foi implementado lógica para adicionar em uma lista singleton os produtos que estão dentro de carrinho de compras. Está em desenvolvimento, ainda falta exibir, agora apenas armazena.

+++ app/src/main/java/dev/edsoncamargo/models/Product.kt
@@ -8,5 +8,5 @@ data class Product(
     val qtdMinEstoque: Int,
     val ativoProduto: Boolean,
     val idProduto: Int,
-    val descontoPromocao: Int
+    val descontoPromocao: Double
 )
